test: split e2e test suite for better performance
