+++ apps/calendar/playwright/configs.ts
@@ -1,0 +1,5 @@
+export const WEEK_VIEW_PAGE_URL =
+  'http://localhost:6006/iframe.html?id=weekview--fixed-events&args=&viewMode=story';
+
+export const MONTH_VIEW_PAGE_URL =
+  'http://localhost:6006/iframe.html?id=monthview--fixed-events&args=&viewMode=story';
