+++ apps/calendar/playwright/month/basic.spec.ts
@@ -1,0 +1,13 @@
+import { expect, test } from '@playwright/test';
+
+import { MONTH_VIEW_PAGE_URL } from '../configs';
+
+test.beforeEach(async ({ page }) => {
+  await page.goto(MONTH_VIEW_PAGE_URL);
+});
+
+test('basic test', async ({ page }) => {
+  const events = await page.$$('.toastui-calendar-weekday-event');
+
+  expect(events).toHaveLength(4);
+});
